Pipline/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pymongo>=4.6.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
langgraph>=0.0.45
langchain>=0.0.350
PyMuPDF>=1.23.9
pandas>=2.1.4
numpy>=1.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=0.24.0
loguru>=0.7.2
httpx>=0.24.0
tenacity>=8.2.0
